all html files ready for bootstrap4
