+++ src/main/java/com/jm/jobseekerplatform/controller/MainController.java
@@ -37,12 +37,6 @@ public class MainController {
     @Autowired
     private EmployerService employerService;
 
-    @Autowired
-    private EmployerProfileService employerProfileService;
-
-    @Autowired
-    private TagService tagService;
-
     private UserRole roleSeeker = new UserRole("ROLE_SEEKER");
 
     @Value("${google.maps.api.key}")
@@ -149,14 +143,5 @@ public class MainController {
         model.addAttribute("logoimg", Base64.getEncoder().encodeToString(vacancy.getEmployerProfile().getLogo()));
 
         return "vacancy";
-    }
-
-    @RequestMapping(value = "/admin/tags", method = RequestMethod.GET)
-    public String UsersViewPage(Model model) {
-
-        List<Tag> tags = tagService.getAll();
-        model.addAttribute("tags", tags);
-
-        return "admin_tags";
     }
 }
