+++ src/main/resources/templates/admin_addUser.html
@@ -15,7 +15,6 @@
         }
     </style>
 </head>
-
 <body>
 
 <div layout:fragment="content">
