+++ src/main/resources/templates/admin_tags.html
@@ -1,90 +1,50 @@
 <!DOCTYPE html>
-<html xmlns:th="http://www.thymeleaf.org" lang="en">
+<html lang="en"
+      xmlns:th="http://www.thymeleaf.org"
+      xmlns:layout="http://www.ultraq.net.nz/thymeleaf/layout"
+      layout:decorator="layouts/adminlayouts" th:with="side='tag'">
 <head>
-    <meta charset="UTF-8">
-    <meta name="_csrf" th:content="${_csrf.token}"/>
-    <meta name="_csrf_header" th:content="${_csrf.headerName}"/>
-    <title>Title</title>
-
-    <!-- Bootstrap core CSS -->
-    <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/css/bootstrap.min.css" integrity="sha384-ggOyR0iXCbMQv3Xipma34MD+dH/1fQ784/j6cY/iJTQUOhcWr7x9JvoRxT2MZw1T" crossorigin="anonymous">
-
+    <title>Tag Page</title>
 </head>
-<body>
 
-<table class="table table-striped">
-    <thead class="" style="border-width: 0px;">
-    <tr style="font-weight: bold; ">
-        <th scope="col">Id</th>
-        <th scope="col">Имя</th>
-        <th scope="col">Утвердить</th>
-        <th scope="col">Удалить</th>
-    </tr>
-    </thead>
-    <tbody>
-    <tr th:each="tag,tagCount: ${tags}" style="font-weight: normal;">
-        <th th:text="${tag.getId()}"></th>
-        <th th:text="${tag.getName()}"></th>
-        <th th:if="${tag.getVerified()}"></th>
-        <th th:unless="${tag.getVerified()}"><button th:onclick="'changeVerifiedTag(' + ${tag.getId()} + ')'">OK</button></th>
-        <th><button th:onclick="'deleteTag(' + ${tag.getId()} + ');'">OK</button>
-        </th>
-    </tr>
-    </tbody>
-</table>
+<div class="container-fluid" layout:fragment="content">
+    <div class="row">
+        <table class="table table-striped" style="margin-left: 20%; margin-right: 20% ">
+            <thead class="" style="border-width: 0px;">
+            <tr style="font-weight: bold; ">
+                <th scope="col">Id</th>
+                <th scope="col">Имя</th>
+                <th scope="col">Утвердить</th>
+                <th scope="col">Удалить</th>
+            </tr>
+            </thead>
+            <tbody>
+            <tr th:each="tag,tagCount: ${tags}" style="font-weight: normal;">
+                <td th:text="${tag.getId()}"></td>
+                <td th:text="${tag.getName()}"></td>
+                <td>
+                    <div th:unless="${tag.getVerified()}">
+                        <button th:onclick="'changeVerifiedTag(' + ${tag.getId()} + ')'" type="button"
+                                class="btn btn-primary"> утвердить
+                        </button>
+                    </div>
+                </td>
+                <td>
+                    <button th:onclick="'deleteTag(' + ${tag.getId()} + ');'" type="button" class="btn btn-primary">
+                        удалить
+                    </button>
+                </td>
+                <!--<th >-->
+                <!---->
+                <!--</th>-->
 
-<script
-        src="https://code.jquery.com/jquery-3.4.1.js"
-        integrity="sha256-WpOohJOqMqqyKL9FccASB9O0KwACQJpFTUBLTYOVvVU="
-        crossorigin="anonymous"></script>
-<script src="https://stackpath.bootstrapcdn.com/bootstrap/4.3.1/js/bootstrap.min.js" integrity="sha384-JjSmVgyd0p3pXB1rRibZUAYoIIy6OrQ6VrjIEaFf/nJGzIxFDsf4x0xIM+B07jRM" crossorigin="anonymous"></script>
-
-<script>
-
-    function deleteTag(userId) {
-
-        var token = $("meta[name='_csrf']").attr("content");
-        var header = $("meta[name='_csrf_header']").attr("content");
-        $.ajax({
-            url: "/api/tags/" + userId,
-            contentType: 'application/json; charset=utf-8',
-            dataType: "json",
-            type: "DELETE",
-            beforeSend: function (request) {
-                request.setRequestHeader(header, token);
-            },
-            success: function () {
-                location.href = '/admin/tags';
-            },
-            error: function (error) {
-                console.log(error);
-                alert(error.toString());
-            }
-        });
-    }
-
-    function changeVerifiedTag(userId) {
-
-        var token = $("meta[name='_csrf']").attr("content");
-        var header = $("meta[name='_csrf_header']").attr("content");
-        $.ajax({
-            contentType: 'application/json; charset=utf-8',
-            url: "/api/tags/change_verified/" + userId,
-            dataType: "json",
-            type: "PUT",
-            beforeSend: function (request) {
-                request.setRequestHeader(header, token);
-            },
-            success: function (data, status, jqXHR) {
-                location.href = '/admin/tags';
-            },
-            error: function (error) {
-                console.log(error);
-                alert(error.toString());
-            }
-        });
-    }
-
-</script>
+            </tr>
+            </tbody>
+        </table>
+    </div>
+</div>
+<div  layout:fragment="content-script">
+    <script type="text/javascript" src="/js/admin/admin.js"></script>
+</div>
 </body>
 </html>
