+++ src/main/resources/templates/seeker.html
@@ -8,8 +8,8 @@
     <script type="text/javascript" src="/js/seeker.js"></script>
 </head>
 <body>
-<div id="sProfileDiv" class="main-block chromeFix" layout:fragment="content">
-    <div class="container-fluid" style="border: 1px solid #7295b1; background: #ced4da">
+<div id="sProfileDiv"  layout:fragment="content">
+    <div class="container-fluid" style="margin-left: 15%;margin-right: 15%">
         <div class="row">
             <div class="col-sm-1">
             </div>
@@ -28,17 +28,18 @@
             </div>
         </div>
         <div class="row">
-            <div class="col-sm-4">
+            <div class="col-sm-1">
             </div>
             <div class="col-sm-6">
                 <span>Обо мне:</span><br/>
                 <div class="well well-sm">
                     <span th:text="${sprofile.description}"></span>
                 </div>
+                <br>
                 <!-- кнопка приглашения -->
                 <button type="button" th:onclick="'showInviteModal('+${sprofile.id}+')'"
-                        class="btn btn-basic"><span
-                        class="glyphicon glyphicon-plus"></span>
+                        class="btn btn-primary">
+                    <i class="fas fa-user-plus"></i>
                     InviteFriend
                 </button>
             </div>
@@ -50,8 +51,8 @@
             <!-- Modal content-->
             <div class="modal-content">
                 <div class="modal-header">
-                    <button type="button" class="close" data-dismiss="modal">&times;</button>
                     <h2 class="modal-title"><span id="PMHeadline"></span></h2>
+                    <button type="button" class="close" data-dismiss="modal">&times;</button>
                 </div>
                 <div class="modal-body">
                     <h4><a id="PMlink"></a></h4>
@@ -60,7 +61,7 @@
                     </div>
                 </div>
                 <div class="modal-footer">
-                    <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
+                    <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
                 </div>
             </div>
 
@@ -72,8 +73,8 @@
             <!-- Modal content-->
             <div class="modal-content">
                 <div class="modal-header">
-                    <button type="button" class="close" data-dismiss="modal">&times;</button>
                     <h2 class="modal-title">Вы можете пригласить друга здесь</h2>
+                    <button type="button" class="close" data-dismiss="modal">&times;</button>
                 </div>
                 <div class="modal-body text-center edit-form">
 
@@ -92,16 +93,16 @@
                 <div class="modal-footer">
                     <button type="submit" onclick="inviteFriend(document.getElementById('seeker-input-add').value,
                     document.getElementById('friend-input-add').value)"
-                            class="btn btn-basic "><span
-                            class="glyphicon glyphicon-envelope"></span>
+                            class="btn btn-primary "><i class="fas fa-envelope"></i>
                         Send Invitation
                     </button>
-                    <button type="button" class="btn btn-default" data-dismiss="modal">Close</button>
+                    <button type="button" class="btn btn-secondary" data-dismiss="modal">Close</button>
                 </div>
             </div>
 
         </div>
     </div>
 </div>
+
 </body>
 </html>
