+++ src/main/resources/templates/vacancy.html
@@ -70,7 +70,7 @@
                 <div style="padding: 15px;">
                     <button type="button" class="btn btn-success"
                             onclick="alert('кнопка не работает')" style="width: 150px;">Откликнуться
-                        <span class="glyphicon glyphicon-envelope"></span>
+                        <i class="fas fa-envelope"></i>
                     </button>
                 </div>
             </div>
